MAKE `TrackActivityBuilderTest` test
