+++ tests/Feature/Builders/TrackTrafficBuilderTest.php
@@ -5,6 +5,7 @@ namespace Sfneal\Tracking\Tests\Feature\Builders;
 
 
 use Sfneal\Tracking\Models\TrackTraffic;
+use Sfneal\Tracking\Tests\Feature\Builders\Traits\WhereUserTests;
 
 class TrackTrafficBuilderTest extends BuilderTestCase
 {
